Security Feature:
Each property of an entity which has been declared public corresponds to the attribute
of the database table. So, if an unknown new property which is not an attribute of the 
table has been set at runtime accidentally by mistake or intensionally, then there might
be an error while updating record or populalting a new record, because that property 
does not exist in the table. So such property has to be removed before adding a new 
record or updating an existing record. Romoving of such unwanted or undefined property 
of an entity class is done by the function removeUndefinedProperty().